Move admin functions and hooks to a separate file in "includes" folder
