+++ wp-postratings.php
@@ -54,17 +54,6 @@ global $wpdb;
 $wpdb->ratings = $wpdb->prefix.'ratings';
 
 
-### Function: Ratings Administration Menu
-add_action('admin_menu', 'ratings_menu');
-function ratings_menu() {
-    add_menu_page(__('Ratings', 'wp-postratings'), __('Ratings', 'wp-postratings'), 'manage_ratings', 'wp-postratings/postratings-manager.php', '', 'dashicons-star-filled');
-
-    add_submenu_page('wp-postratings/postratings-manager.php', __('Manage Ratings', 'wp-postratings'), __('Manage Ratings', 'wp-postratings'), 'manage_ratings', 'wp-postratings/postratings-manager.php');
-    add_submenu_page('wp-postratings/postratings-manager.php', __('Ratings Options', 'wp-postratings'), __('Ratings Options', 'wp-postratings'),  'manage_ratings', 'wp-postratings/postratings-options.php');
-    add_submenu_page('wp-postratings/postratings-manager.php', __('Ratings Templates', 'wp-postratings'), __('Ratings Templates', 'wp-postratings'),  'manage_ratings', 'wp-postratings/postratings-templates.php');
-}
-
-
 ### Function: Display The Rating For The Post
 function the_ratings($start_tag = 'div', $custom_id = 0, $display = true) {
     global $id;
@@ -768,37 +757,6 @@ function ratings_sorting($local_wp_query) {
 }
 
 
-### Function Show Ratings Column in WP-Admin
-add_action('manage_posts_custom_column', 'add_postratings_column_content');
-add_filter('manage_posts_columns', 'add_postratings_column');
-add_action('manage_pages_custom_column', 'add_postratings_column_content');
-add_filter('manage_pages_columns', 'add_postratings_column');
-function add_postratings_column($defaults) {
-    $defaults['ratings'] = 'Ratings';
-    return $defaults;
-}
-
-
-### Functions Fill In The Ratings
-function add_postratings_column_content($column_name) {
-    global $post;
-    if($column_name == 'ratings') {
-        if(function_exists('the_ratings')) {
-            $template = str_replace('%RATINGS_IMAGES_VOTE%', '%RATINGS_IMAGES%<br />', stripslashes(get_option('postratings_template_vote')));
-            echo expand_ratings_template($template, $post, null, 0, false);
-        }
-    }
-}
-
-
-### Function Sort Columns
-add_filter('manage_edit-post_sortable_columns', 'sort_postratings_column');
-add_filter('manage_edit-page_sortable_columns', 'sort_postratings_column');
-function sort_postratings_column($defaults)
-{
-    $defaults['ratings'] = 'ratings';
-    return $defaults;
-}
 add_action('pre_get_posts', 'sort_postratings');
 function sort_postratings($query) {
     if(!is_admin())
@@ -1223,4 +1181,5 @@ require_once('includes/scripts.php');
 require_once('includes/shortcodes.php');
 require_once('includes/widgets.php');
 require_once('includes/activation.php');
+require_once('includes/admin.php');
 require_once('postratings-stats.php');
